Refactored Contact details of the project
