+++ src/features/projects/screens/SingleProjectDetails.tsx
@@ -3,15 +3,10 @@ import { Elements } from '@stripe/react-stripe-js';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import React, { ReactElement } from 'react';
-import LazyLoad from 'react-lazyload';
 import ReactPlayer from 'react-player/lazy';
 import ReadMoreReact from 'read-more-react';
 import Sugar from 'sugar';
 import BackButton from '../../../../public/assets/images/icons/BackButton';
-import BlackTree from '../../../../public/assets/images/icons/project/BlackTree';
-import Email from '../../../../public/assets/images/icons/project/Email';
-import Location from '../../../../public/assets/images/icons/project/Location';
-import WorldWeb from '../../../../public/assets/images/icons/project/WorldWeb';
 import getImageUrl from '../../../utils/getImageURL';
 import getStripe from '../../../utils/stripe/getStripe';
 import { ThemeContext } from '../../../theme/themeContext';
@@ -55,46 +50,6 @@ function SingleProjectDetails({ project }: Props): ReactElement {
     ? getImageUrl('project', 'large', project.image)
     : '';
 
-  const contactAddress = project.tpo && project.tpo.address
-    ? (project.tpo.address.address ? project.tpo.address.address + ', ' : '')
-    + (project.tpo.address.city ? project.tpo.address.city + ', ' : '')
-    + (project.tpo.address.zipCode ? project.tpo.address.zipCode + ' ' : '')
-    + (project.tpo.address.country ? t('country:' + project.tpo.address.country.toLowerCase()) : '')
-    : t('donate:unavailable');
-    const projectWebsiteLink = project.website ? project.website.includes("http") || project.website.includes("https") ? project.website : `http://${project.website}` : t('donate:unavailable')
-
-  const contactDetails = [
-    {
-      id: 1,
-      icon: <BlackTree color={styles.highlightBackground} />,
-      text: t('donate:viewProfile'),
-      link: project.tpo.slug,
-    },
-    {
-      id: 2,
-      icon: <WorldWeb color={styles.highlightBackground} />,
-      text: project.website ? project.website : t('donate:unavailable'),
-      link: projectWebsiteLink,
-    },
-    {
-      id: 3,
-      icon: <Location color={styles.highlightBackground} />,
-      text: contactAddress,
-      link: project.coordinates
-        ? `https://maps.google.com/?q=${contactAddress}`
-        : null,
-    },
-    {
-      id: 4,
-      icon: <Email color={styles.highlightBackground} />,
-      text:
-        project.tpo && project.tpo.email
-          ? project.tpo.email
-          : t('donate:unavailable'),
-      link:
-        project.tpo && project.tpo.email ? `mailto:${project.tpo.email}` : null,
-    },
-  ];
   React.useEffect(() => {
     const code = window.localStorage.getItem('countryCode') || 'DE';
     setCountryCode(code);
@@ -306,9 +261,8 @@ function SingleProjectDetails({ project }: Props): ReactElement {
                     {species ? <PlantSpecies species={species} /> : null }
                     {co2 ? (<CarbonCaptured co2={co2} />) : null} */}
 
-              {contactDetails ? (
-                <ProjectContactDetails contactDetails={contactDetails} />
-              ) : null}
+                <ProjectContactDetails project={project} />
+             
             </div>
           </div>
         </div>
